pyproject.toml
[project]
name = "bchwtz-gateway-microservices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio==3.4.3",
    "attr==0.3.2",
    "attrs==25.1.0",
    "bleak==0.22.3",
    "crcmod==1.7",
    "fastjsonschema==2.21.1",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-core==5.7.2",
    "nbformat==5.10.4",
    "paho-mqtt==2.1.0",
    "platformdirs==4.3.6",
    "python-dotenv==1.0.1",
    "reactivex==4.0.4",
    "referencing==0.36.2",
    "rpds-py==0.23.1",
    "ruuvitag-sensor==3.1.0",
    "termcolor==2.5.0",
    "traitlets==5.14.3",
    "typing-extensions==4.12.2",
]
